requirements.txt
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.3
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
pymongo==4.2.0
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.1
